+++ docs/conf.py
@@ -86,6 +86,6 @@ autodoc_pydantic_field_list_validators = False
 
 intersphinx_mapping = {
     "python": ("https://docs.python.org/3", None),
-    "qiskit": ("https://docs.quantum.ibm.com/api/qiskit/1.4", None),
+    "qiskit": ("https://quantum.cloud.ibm.com/docs/api/qiskit/1.4", None),
     "pydantic": ("https://docs.pydantic.dev/latest/", None),
 }