+++ CHANGELOG.md
@@ -55,7 +55,7 @@
 
 ## qiskit-aqt-provider v1.3.0
 
-* Point Qiskit docs links to docs.quantum.ibm.com (#135)
+* Point Qiskit docs links to quantum.cloud.ibm.com/docs (#135)
 * Remove references to the deprecated function `qiskit.execute` (#136)
 * Pin Qiskit dependency strictly below 1.0 (#137)
 * Remove the Grover 3-SAT example (#137)